feat: mgmt command queue task for logs to splunk
